requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Login==0.6.2
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
python-dotenv==1.0.0
Pillow==10.1.0
pandas==2.1.4
openpyxl==3.1.2
firebase-admin>=6.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0